move binding logic
